# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: A new user has been added
        "400":
          description: Wrong input parameters
  /Login:
    post:
      tags:
        - Auth
      summary: Login an exist user
      description: An exist user login to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
      responses:
        "200":
          description: >
            User was found in the DB and authenticated successfully.
            The session ID is returned in a cookie named `session`. 
            You should include this cookie in subsequent requests
          # headers:
          #   Set-Cookie:
          #     schema:
          #       type: string
          #       example: session=87s97sdhgf279293f8w0e9
            # Note that the Set-Cookie header and securitySchemes are not connected in any way, and the Set-Header definition is for documentation purposes only.
        "401":
          description: Authentication failed
  /logout:
    post:
      tags:
        - Auth
      summary: Logout a user from the system and delete his cookie
      responses:
        "200":
          description: "The user has ben successfully logged out"
        "400":
          description: Bad request
...
  /league/getDetails:
    get:
      tags:
        - League
      summary: Return the details about the current league
      responses:
        "200":
          description: Data about the current league returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  leagueDetails:
                    $ref: "#/components/schemas/League"
                  userFavoriteMatches:
                    type: array
                    description: For logged in user - maximum 3 Favorite Matches
                    items:
                      $ref: "#/components/schemas/FutureMatchInformation"
        "400":
          description: Wrong input parameters
  /league/search/{Search_Query}:
    get:
      tags: 
        - League
      summary: Return search results
      description: Returns all results according to the user's search query - from partial or full query 
      parameters: 
        - name: Search_Query
          in: path
          description: User's search query
          required: true
          schema:
            type: string
            example: vic
        - name: Search_Type
          in: query
          description: Search on teams or players 
          required: true
          style: form
          schema:
            type: array
            items:
              type: string
              enum:
              - Players
              - Teams
        - name: Sort_Teams_Alphabetical
          in: query
          description: Sort the teams by names in alphabetical order 
          style: form
          schema:
            type: array
            items:
              type: string
              enum: 
              - no
              - yes
        - name: Sort_Players
          in: query
          description: Sort the players
          style: form
          schema:
            type: array
            items:
              type: string
              enum: 
              - no
              - yes
        - name: Sort_Players_By
          in: query
          description: Sort the players by name or team name
          style: form
          schema:
            type: array
            items:
              type: string
              enum: 
              - own name
              - team name
        - name: Filter_Players
          in: query
          description: Filter the players/ by position or team name
          schema:
            type: string
      responses:
        '200':
          description: Search results returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:   
                    type: array
                    items:
                      $ref: '#/components/schemas/PlayerPreview'
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        teamName:
                          type: string
                        teamLogo:
                          type: string
                          format: URL
                      example:
                        teamName: Beitar Jerusalem
                        teamLogo: 'https://cdn.sportmonks.com/images//soccer/teams/11/939.png'
        '400':
          description: Bad request
        '204':
          description: No Content
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username contains between 3 to 8 letters only
          example: ladygaga
        firstname:
          type: string
          description: user's first name
          example: Stefani
        lastname:
          type: string
          description: user's last name
          example: Germanotta
        country:
          type: string
          description: user's country
          example: USA
        password:
          maxLength: 10
          minLength: 5
          type: string
          description:
            password contains between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: lady@56
        email:
          pattern: email
          type: string
          description: user's email
          example: ladygaga@gmail.com
        image_url:
          pattern: URL
          type: string
          description: URL for the user's image
          example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          description: username contains letters only
          example: naorbe
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          description: password contains at least one number and one special char
          example: naor@55
    League:
      type: object
      properties:
        league_name:
          type: string
          description: league name
          example: Superliga
        current_season_name:
          type: string
          description: current season name
          example: "2020/2021"
        current_stage_name:
          type: string
          description: current stage name
          example: Championship Round
        first_next_match:
          $ref: '#/components/schemas/FutureMatchInformation'
    PlayerPreview:
      type: object
      properties:
        playerID:
          type: number
          description: Player's ID
          example: 1 
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
        position:
          type: integer
          description: player's position in team
          example: 1
        team_name:
          type: string
          description: player's team name
          example: København
...
    FutureMatchInformation:
      type: object
      description: The information about future match
      required:
        - matchDate
        - localTeamName
        - visitorTeamName
        - venueName
      properties:
        matchID:
          type: number
          description: Match's ID
          example: 1 
        matchDate:
          type: string
          format: date-time
          description: Match's date and time of beginning
          example: "2021-07-02 12:00:00"
        localTeamName:
          type: string
          description: The local team name
          example: Viborg
        visitorTeamName:
          type: string
          description: The visitor team name
          example: Silkeborg
        venueName:
          type: string
          description: The match stadium name
          example: Sydbank Park
        refereeInformation:
          $ref: '#/components/schemas/Referee'
...
    Referee:
      type: object
      description: Referee information
      properties:
        refereeID:
          type: number
          description: Referee's ID
          example: 10
        firstname:
          type: string
          description: Referee's first name
          example: Stefani
        lastname:
          type: string
          description: Referee's last name
          example: Germanotta
        course:
          type: string
          description: Referee's course
          example: Regular
          enum: 
          - Regular
          - Main